+++ boot/stage2/disk.h
@@ -14,7 +14,7 @@ typedef struct {
 
 #define SUCCESSFUL 0
 #define FAILED 1
-void linear_to_segment_offset(uint32_t linear_address, uint16_t *segment_out,
+void linear_to_segment_offset(uintptr_t linear_address, uint16_t *segment_out,
                               uint16_t *offset_out);
 
 void lba_to_chs(uint32_t lba, uint8_t *cylinder_out, uint8_t *head_out,
@@ -22,5 +22,5 @@ void lba_to_chs(uint32_t lba, uint8_t *cylinder_out, uint8_t *head_out,
                 uint8_t sectors_per_track);
 
 bool disk_read(uint8_t drive, uint32_t lba, uint16_t num_sectors,
-               uint32_t storage_address, uint8_t heads_per_cylinder,
+               uintptr_t storage_address, uint8_t heads_per_cylinder,
                uint8_t sectors_per_track);
